ajout d'un message d'avertissement avant supression de commentaire ou billet

+++ app/view/backend/backendView.php
@@ -43,7 +43,7 @@
 
         <div class="row mt-5">
           <a href="createNewPost" class="btn btn-primary mt-1 col-lg-3 col-md-6 col-sm-12">Ajouter un billet</a>
-          <a href="#reportList" class="btn btn-danger mt-1 ml-2 col-lg-3 col-md-6 col-sm-12">messages signalés</a>
+          <a href="#reportList" class="btn btn-danger mt-1 ml-lg-2 ml-sm-0 ml-md-0 col-lg-3 col-md-6 col-sm-12">messages signalés</a>
         </div>
         <div class="row">
           <h2 class="mt-4">Billets : </h2>
@@ -60,7 +60,7 @@
               </div>
               <div class=" col-md-3 d-flex justify-content-md-end">
                 <button class="btn btn-warning p-2 "><a href="editPost&postId=<?= $post->id();?>"><i class="fa fa-edit text-white"></i> </a></button>
-                <button class="btn btn-danger p-2 ml-2"><a href="deletePost&postId=<?= $post->id();?>"><i class="fa fa-trash text-white"></i></a></button>
+                <button class="btn btn-danger p-2 ml-2"><a href="deletePost&postId=<?= $post->id();?>" OnClick="return confirm('Voulez-vous vraiment supprimer ce billet ?');"><i class="fa fa-trash text-white"></i></a></button>
               </div>
             </div>
           </li>
@@ -83,7 +83,7 @@
               </div>
               <div class="col-md-3 d-flex justify-content-md-end">
                 <button class="btn btn-success p-2"><a href="authorizedComment&commentId=<?= $report->id();?>"><i class="fa fa-check text-white"></i></a></button>
-                <button class="btn btn-danger p-2 ml-2"><a href="deleteComment&commentId=<?= $report->id();?>"><i class="fa fa-trash text-white"></i></a></button>
+                <button class="btn btn-danger p-2 ml-2"><a href="deleteComment&commentId=<?= $report->id();?>" OnClick="return confirm('Voulez-vous vraiment supprimer ce commentaire ?');"><i class="fa fa-trash text-white"></i></a></button>
               </div>
             </div>
           </li>
